+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 - 2022-12-07
+
+### Fixed
+
+- fixed route matching with `matcher.query` (was causing a runtime error when used with relative URLs)
+- fixed route matching with relative `url` (string matcher)
+
 ## 0.2.0 - 2022-12-06
 
 ### Added